% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather_icons.R
\name{feather_icons}
\alias{feather_icons}
\title{A Feather Icon div}
\usage{
feather_icons(icon)
}
\arguments{
\item{icon}{the icon}

\item{...}{optional content to the div}
}
\value{
an HTML div
}
\description{
A Feather Icon div
}
\examples{
feather_icons("circle", "Hello circle!")
}
