% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\docType{data}
\name{icons_list}
\alias{icons_list}
\title{List of feather icons}
\format{A vector of 266 icons}
\source{
\url{https://feathericons.com/}
}
\usage{
icons_list
}
\description{
Simply beautiful open source icons
}
\keyword{datasets}
